Add API token authentication scheme (#3)
* Create new authentication models and migrations

This commit introduces the Django models and their respective database migrations for the API token authentication. Existing misconfiguration of the project is patched and tested to accommodate for the new models.

The `KeyCloakClient` is updated to properly link internal user models to Keycloak users

* Remove redundant FlotoAuthBackend

The FlotoAuthBackend class was removed because it was not adding any new functionality. It was merely wrapping the get_or_create_user method of the parent OIDCAuthenticationBackend class without additional logic, which made it functionally redundant.

* Add test for API token authentication

* Add interface for users to obtain API keys

Removes noisy demo text from user page, and adds an interface to reveal and copy the FLOTO API key.

Exposes the user page via the dashboard header.